tidy up script

+++ HoraceExampleScriptsLink.sh
@@ -1,1 +1,7 @@
-ln -s /home/resources/HoraceExampleScripts/ ~/Desktop/HoraceExampleScripts
+if [ ! -f  ~/Desktop/HoraceExampleScripts];
+then
+    exit 0
+else
+    ln -s /home/resources/HoraceExampleScripts/ ~/Desktop/HoraceExampleScripts
+fi
+}
